+++ test/ui-test/lightspeedUiTestExplorerTest.ts
@@ -1,4 +1,5 @@
 // BEFORE: ansible.lightspeed.enabled: true
+
 import {
   ActionsControl,
   ActivityBar,
@@ -8,42 +9,30 @@ import {
   InputBox,
   ModalDialog,
   SideBarView,
+  VSBrowser,
   ViewControl,
   ViewSection,
-  VSBrowser,
-  until,
-  WebElement,
   WebView,
   WebviewView,
   Workbench,
 } from "vscode-extension-tester";
 import {
-  expectNotification,
   getFixturePath,
   getModalDialogAndMessage,
   getWebviewByLocator,
   sleep,
+  workbenchExecuteCommand,
 } from "./uiTestHelper";
-import { Key } from "selenium-webdriver";
+import { WizardGenerationActionType } from "../../src/definitions/lightspeed";
+import { PlaybookGenerationActionEvent } from "../../src/interfaces/lightspeed";
 import { expect } from "chai";
 import axios from "axios";
 
-const trialNotificationMessage =
-  "Ansible Lightspeed is not configured for your organization, click here to start a 90-day trial.";
-
-before(function () {
-  if (process.platform !== "darwin") {
-    this.skip();
-  }
-});
-
-describe("Test One Click Trial feature", () => {
+describe("Test Lightspeed Explorer features", () => {
   let workbench: Workbench;
   let explorerView: WebviewView;
   let modalDialog: ModalDialog;
   let dialogMessage: string;
-  let playbookGeneration: WebView;
-  let submitButton: WebElement;
   let sideBar: SideBarView;
   let view: ViewControl;
   let adtView: ViewSection;
@@ -76,7 +65,7 @@ describe("Test One Click Trial feature", () => {
     // Set "UI Test" and "One Click" options for mock server
     await axios.post(
       `${process.env.TEST_LIGHTSPEED_URL}/__debug__/options`,
-      ["--ui-test", "--one-click"],
+      ["--ui-test"],
       { headers: { "Content-Type": "application/json" } },
     );
   });
@@ -146,130 +135,87 @@ describe("Test One Click Trial feature", () => {
     await sleep(2000);
   });
 
-  it("Verify Ansible Lightspeed webview now contains user's information", async () => {
-    await explorerView.switchToFrame(5000);
-    const div = await explorerView.findWebElement(
-      By.id("lightspeedExplorerView"),
+  it("Verify Ansible Lightspeed Generate Playbook button", async () => {
+    await explorerView.switchToFrame(2000);
+
+    const generatePlaybookButton = await explorerView.findWebElement(
+      By.id("lightspeed-explorer-playbook-generation-submit"),
     );
-    const text = await div.getText();
-    expect(text).contains("Logged in as: ONE_CLICK_USER (unlicensed)");
-    await explorerView.switchBack();
-    await expectNotification("Welcome back ONE_CLICK_USER (unlicensed)");
-  });
+    expect(generatePlaybookButton).not.to.be.undefined;
 
-  it("Verify the Refresh icon is found on the title of Explorer view", async () => {
-    const refreshIcon = await explorerView.findWebElement(
-      By.xpath("//a[contains(@class, 'codicon-refresh')]"),
-    );
-    expect(refreshIcon, "refreshIcon should not be undefined").not.to.be
-      .undefined;
-
-    // Set "UI Test", One Click" and "Me Uppercase" options for mock server
-    await axios.post(
-      `${process.env.TEST_LIGHTSPEED_URL}/__debug__/options`,
-      ["--ui-test", "--one-click", "--me-uppercase"],
-      { headers: { "Content-Type": "application/json" } },
-    );
-
-    await refreshIcon.click(); // make sure if it could be clicked
-
+    // Open playbook generation webview.
+    await generatePlaybookButton.click();
     await sleep(2000);
-    await explorerView.switchToFrame(5000);
-    const div = await explorerView.findWebElement(
-      By.id("lightspeedExplorerView"),
-    );
-    const text = await div.getText();
-    expect(text).contains("LOGGED IN AS: ONE_CLICK_USER (UNLICENSED)");
-    await explorerView.switchBack();
-  });
 
-  it("Invoke Playbook generation", async () => {
-    await workbench.executeCommand("Ansible Lightspeed: Playbook generation");
-    await sleep(2000);
-    playbookGeneration = await getWebviewByLocator(
+    // Locate the playbook explanation webview
+    await explorerView.switchBack();
+    const webView = await getWebviewByLocator(
       By.xpath("//*[text()='Create a playbook with Ansible Lightspeed']"),
     );
+    expect(webView, "webView should not be undefined").not.to.be.undefined;
 
-    // Set input text and invoke summaries API
-    const textArea = await playbookGeneration.findWebElement(
-      By.xpath("//vscode-text-area"),
+    await webView.switchBack();
+    await workbenchExecuteCommand("View: Close All Editor Groups");
+
+    /* verify generated events */
+    const expected = [
+      [WizardGenerationActionType.OPEN, undefined, 1],
+      [WizardGenerationActionType.CLOSE_CANCEL, 1, undefined],
+    ];
+    const res = await axios.get(
+      `${process.env.TEST_LIGHTSPEED_URL}/__debug__/feedbacks`,
     );
-    expect(textArea, "textArea should not be undefined").not.to.be.undefined;
-    submitButton = await playbookGeneration.findWebElement(
-      By.xpath("//vscode-button[@id='submit-button']"),
-    );
-    expect(submitButton, "submitButton should not be undefined").not.to.be
-      .undefined;
-    //
-    // Note: Following line should succeed, but fails for some unknown reasons.
-    //
-    // expect((await submitButton.isEnabled()), "submit button should be disabled by default").is.false;
-    await textArea.sendKeys("Create an azure network.");
-    expect(
-      await submitButton.isEnabled(),
-      "submit button should be enabled now",
-    ).to.be.true;
-    await submitButton.click();
-    await playbookGeneration.switchBack();
-    await sleep(2000);
-    await expectNotification(
-      trialNotificationMessage,
-      true, // click button
-    );
-    await new EditorView().closeAllEditors();
+    expect(res.data.feedbacks.length).equals(expected.length);
+    for (let i = 0; i < expected.length; i++) {
+      const evt: PlaybookGenerationActionEvent =
+        res.data.feedbacks[i].playbookGenerationAction;
+      expect(evt.action).equals(expected[i][0]);
+      expect(evt.fromPage).equals(expected[i][1]);
+      expect(evt.toPage).equals(expected[i][2]);
+    }
   });
 
-  it("Invoke Playbook explanation", async () => {
+  it("Verify Ansible Lightspeed Explain Playbook button", async () => {
     const folder = "lightspeed";
     const file = "playbook_4.yml";
     const filePath = getFixturePath(folder, file);
 
     // Open file in the editor
     await VSBrowser.instance.openResources(filePath);
-    await sleep(1000);
+
+    await explorerView.switchToFrame(2000);
+
+    const explainPlaybookButton = await explorerView.findWebElement(
+      By.id("lightspeed-explorer-playbook-explanation-submit"),
+    );
+    expect(explainPlaybookButton).not.to.be.undefined;
 
     // Open playbook explanation webview.
-    await workbench.executeCommand(
-      "Explain the playbook with Ansible Lightspeed",
-    );
+    await explainPlaybookButton.click();
     await sleep(2000);
-    await expectNotification(
-      trialNotificationMessage,
-      true, // click button
+
+    // Locate the playbook explanation webview
+    await explorerView.switchBack();
+    let webView = (await new EditorView().openEditor(
+      "Explanation",
+      1,
+    )) as WebView;
+    expect(webView, "webView should not be undefined").not.to.be.undefined;
+    webView = await getWebviewByLocator(
+      By.xpath("//div[@class='playbookGeneration']"),
     );
-    await new EditorView().closeAllEditors();
-  });
-
-  it("Invoke Completion", async () => {
-    const folder = "lightspeed";
-    const file = "playbook_3.yml";
-    const filePath = getFixturePath(folder, file);
-    await VSBrowser.instance.openResources(filePath);
-    await sleep(1000);
-
-    const editorView = new EditorView();
-    const tab = await editorView.getTabByTitle(file);
-    await tab.select();
-
-    // Trigger completions API. First space key and backspace look redundant,
-    // but just sending page down and 4 spaces did not work...
-    await tab.sendKeys(" ", Key.BACK_SPACE, Key.PAGE_DOWN, "    ");
-
-    await sleep(4000);
-    await expectNotification(
-      trialNotificationMessage,
-      true, // click button
+    await webView.findWebElement(
+      By.xpath("//h2[contains(text(), 'Playbook Overview and Structure')]"),
     );
 
-    // Revert changes made
-    await tab.select();
-    await tab.sendKeys(Key.CONTROL, "z", "z", "z", Key.NULL);
-    await editorView.closeAllEditors();
+    await webView.switchBack();
+    await workbenchExecuteCommand("View: Close All Editor Groups");
 
-    // The undo's don't seem to (always) work.. So discard changes
-    const dialog = new ModalDialog();
-    await dialog.pushButton(`Don't Save`);
-    await dialog.getDriver().wait(until.stalenessOf(dialog), 2000);
+    /* verify generated events */
+    const res = await axios.get(
+      `${process.env.TEST_LIGHTSPEED_URL}/__debug__/feedbacks`,
+    );
+    expect(res.data.feedbacks.length).equals(1);
   });
 
   it("Sign out using Accounts global action", async () => {
@@ -286,7 +232,7 @@ describe("Test One Click Trial feature", () => {
     expect(menu).not.to.be.undefined;
     if (menu) {
       await menu.select(
-        "ONE_CLICK_USER (unlicensed) (Ansible Lightspeed)",
+        "EXTERNAL_USERNAME (licensed) (Ansible Lightspeed)",
         "Sign Out",
       );
     }
